requirements.txt
flask
langchain-groq
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
faiss-cpu
PyPDF2
scikit-learn
python-dotenv
numpy
matplotlib
opencv-python
tiktoken
easyocr
pdf2image
pyyaml
pypdf
waitress
